Feat: Add update rows with lodash
