+++ package.json
@@ -14,6 +14,7 @@
     "@quasar/extras": "^1.0.0",
     "axios": "^0.21.1",
     "core-js": "^3.6.5",
+    "lodash": "^4.17.21",
     "quasar": "^2.12.0",
     "vue": "^3.0.0",
     "vue-router": "^4.0.0",
