+++ src/router/routes.js
@@ -5,11 +5,9 @@ const routes = [
     children: [{ path: "", component: () => import("pages/Index.vue") }],
   },
   {
-    path: "/operators",
+    path: "/plans",
     component: () => import("layouts/MainLayout.vue"),
-    children: [
-      { path: "", component: () => import("pages/OperatorsPage.vue") },
-    ],
+    children: [{ path: "", component: () => import("pages/PlansPage.vue") }],
   },
 
   // Always leave this as last one,
