--- src/store/module-example/actions.js
@@ -1,1 +1,0 @@
-export function someAction(/* context */) {}
