+++ src/store/plans/mutations.js
@@ -1,0 +1,3 @@
+export function setPlans(state, { plansList }) {
+  state.plansList = plansList;
+}
